+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 * pool: add `RelayPool::disconnect_relay` method ([Yuki Kishimoto])
 * bindings(nostr): expose `as_pretty_json` for some structs ([Yuki Kishimoto])
 * bindings(sdk): expose `Client::fetch_metadata` ([Yuki Kishimoto])
+* bindings(sdk): expose `Client::pool` method ([Yuki Kishimoto])
 * ffi(nostr): expose `Kind::is_*` methods ([Yuki Kishimoto])
 * ffi(sdk): expose `MockRelay` ([Yuki Kishimoto])
 * js(nostr): add `Kind` object ([Yuki Kishimoto])